requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
apturl==0.5.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
bitarray==2.9.2
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.3
cbor2==5.6.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
crytic-compile==0.3.6
cupshelpers==1.0
cytoolz==0.12.3
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duckduckgo_search==6.1.9
duplicity==0.8.21
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
exceptiongroup==1.2.1
fasteners==0.14.1
filelock==3.15.4
Flask==3.0.1
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lru-dict==1.2.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
packaging==24.0
pandas==2.2.0
pandoc==2.3
paramiko==2.9.3
parsimonious==0.9.0
pexpect==4.8.0
phidata==2.4.25
Pillow==9.0.1
plumbum==1.8.2
ply==3.11
prettytable==3.10.0
protobuf==5.26.0
ptyprocess==0.7.0
pyarrow==16.1.0
pycairo==1.20.1
pycryptodome==3.20.0
pycups==2.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyreqwest_impersonate==0.4.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyunormalize==15.1.0
pyxdg==0.27
PyYAML==5.4.1
referencing==0.34.0
regex==2023.12.25
reportlab==3.6.8
requests==2.32.3
rich==13.7.1
rlp==4.0.0
rpds-py==0.18.0
scikit-learn==1.2.2
scipy==1.12.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
slither-analyzer==0.10.1
smmap==5.0.1
sniffio==1.3.1
solc-select==1.0.4
streamlit==1.36.0
systemd-python==234
tenacity==8.4.2
threadpoolctl==3.2.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.10.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
